Rewrote the application help index.
a) make use of ksycoca to speed things up
b) only show entries that actually have a help file

I have heared rumors that the translators decided also to
translate the filenames of the help files. Note that this
will break the current helpcenter completely, so better try
only with english right now.

svn path=/trunk/kdebase/khelpcenter/; revision=50413