Add Xapian support for indexing and searching
Start using Xapian for indexing and searching the documentation: it
works much better than htp://dig, it provides a C++ API, and it is well
maintained.

Introduce two helper tools to create/update an index, and search on it,
starting from the cache.bz2 files generated by meinproc: this way it is
possible to index the whole text, with the entities already expanded
(something which would not be there when parsing the docbook files
directly). Make use of libxml2 to parse the HTML documents, and get all
the text in them.

Xapian and libxml2 are currently considered as mandatory dependencies:
they are portable, generally available everywhere, and not requiring
themselves too many extra dependencies.

CCBUG: 93664
CCBUG: 103266
CCBUG: 115935
CCBUG: 122437
CCBUG: 125276
BUG: 126710
CCBUG: 152671
CCBUG: 158633
CCBUG: 209415
CCBUG: 256397
CCBUG: 266290