Automatically update indexes at startup
Currently the user is asked to manually update the search index, which
is not a friendly solution. Instead, start the indexing update right
after the startup, with a indefinite progress bar showing the on-going
process. Considering the Xapian indexing is generally fast, this won't
cause a big delay between the startup and when the user can start
searching.

Move the queue-based index logic from IndexBuilder to Navigator
directly, getting rid of the separate khc_indexbuilder tool (no need to
have an helper tool to spawn the actual indexing/searcing tools) and
the D-Bus interface to support it, and removing the indexing dialog
(now everything that can be indexed is updated).

This also removes the need for indexers to create <identifier>.exists
files in the index directories.

BUG: 101876
BUG: 164308
BUG: 205126
CCBUG: 222494
CCBUG: 269260