Sync View::langLookup with kio_help
View::langLookup is basically a copy (with a small difference w.r.t.
the .html extension check) of the same function in kio_help, used to
locate a documentation among the installed ones.

While using QLocale::uiLanguages() is not the best solution ever
(kio_help has the same issue, which will be hopefully solved in
kio >= 5.21), at least this avoids to pick the first translation
available alphabetically, instead of the one available for the current
language (or English).