Fix arguments usage logic
khelpcenter can use only a single argument specified on command line:
thus it is useless to open all the arguments, since only the last one
will be displayed at the end.

Bring back the logic as it was before the Frameworks porting, i.e. use
only the first argument, if available, opening the eventually read URL.
A null URL causes the loading of the start page, which is better than
an empty view.