Refresh HTML templates for start and glossary entries pages
The current HTML templates used to show the start page and the pages
of glossary entryes refer to an old KDE style, and thus rendering
almost as if they were not styled at all.  Refresh them applying the
style used e.g. in kio_man/kio_info.

However, two bugs in kio_help prevent the correct loading of the
stylesheets in KHTML:
- 3642b5284e65d94db22a86d52e710dd60e15b230 (fixed in kio >= 5.19)
- 03924e35ee8637be528d746568b4388e8b65a608 (fixed in kio >= 5.20)
since the result when using an older kio is mostly close to the
previous templates (no styles at all), this addition does not cause
regressions.