Fix NavigatorItem::setExpanded conflict
QTreeWidgetItem::setExpanded is not virtual, so the reimplementation in
NavigatorItem will not be called when acting on a QTreeWidgetItem*
(and not on NavigatorItem*).
To make sure items are always notified when they are expanded/collapsed,
connect to the signals emitted by QTreeWidget, and call itemExpanded
(the former setExpanded implementation).  Feels like a kludge though.