Simplify internal scrollkeeper creation
Merge scrollkeeper.desktop into the .directory of the Scrollkeeper
directory, since a single item for it is enough: this allows us to
handle the scrollkeeper tree creation much like other "special" trees.

Also remove in ScrollKeeperTreeBuilder the need for the "after" item,
since all the subtrees can added to the right top-level item.