Switch to search widget only on search
Do not switch on the search widget when the scope (set of enabled
search plugins) is changed or when something is typed in the search
text box. Only switch when the search button is pushed.

As there are some search providers enabled by default now, the search
button should be enabled by default (unless the user changed the
settings later), so even if the search pane is not visible the
default scenario (type something, push search) should work.

BUG: 363830
FIXED-IN: 5.7.1