Clean the status bar message from the HTML part
The message for the status bar that is returned by the HTML part
is considered rich text as it starts by "<qt>", but without real
HTML tags/rich text markers. So clean it before showing it (which
means the removal of the "<qt>" prefix).
Thanks Pino for the QTextDocument hint.

FIXED-IN: 5.7.1
BUG: 361932