Adds missing dependency
Summary: KF5::CoreAddons files are included by khc_xapianindexer, but not declared as dependency. For operating systems like Nixos that is a problem, because every dependency is located in a different path and so, the CoreAddons files are not found in the include path. By adding CoreAddons as target link library, the CoreAddon paths are added to the include paths.

Reviewers: ltoscano

Reviewed By: ltoscano

Subscribers: #documentation

Tags: #documentation

Differential Revision: https://phabricator.kde.org/D7413