Remove duplicate submenus.
Summary:
Signed-off-by: Pavel Mars <mars@basealt.ru>

Remove duplicate submenus.

On the left side of the khelpcenter window is the content ("Contents").
If you open the item "Application Manuals" -> "System" -> "Info Center", the item "Info Center" will generate a tree of additional menu items.

If you open the menu item "KInfoCenter Modules" in the same content and then open the menu items that appear (for example, "Device Viewer", "Memory", "OpenGL" or any other item), the same additional menu item tree will be generated.

This change removes duplication of generated trees.

This is achieved by registering the URLs of the elements of the content tree in a separate container and checking the URL to be opened: if the base address from the URL is registered, then we request the generation of the menu tree from the registered element of the tree, but not from the current one.

BUG: 399274

Reviewers: yurchor, ltoscano

Reviewed By: yurchor, ltoscano

Subscribers: ltoscano, kde-doc-english

Tags: #documentation

Differential Revision: https://phabricator.kde.org/D16308