Fix opening URL's to help files
Summary:
Signed-off-by: Pavel Mars <mars@basealt.ru>

When you select the menu item "KInfoCenter Modules" -> "DMA-Channels" (or another item: Memory, OpenGL, etc.) -> "KInfoCenter" -> "Start KInfoCenter", the help on the right side of the program screen is incorrect.

When you select any of the submenus that appear, the same help is displayed (../index.html).
This is due to the incorrect formation of links to help.
To fix this bug, add "QUrl :: RemoveFragment" to the string "mToc-> setApplication (url.toString (QUrl :: RemoveScheme | QUrl :: RemoveFilename | QUrl :: StripTrailingSlash))" in the navigatorappitem.cpp file.
This expression removes the fragment from the generated link and allows you to properly load the help file.

Reviewers: yurchor, ltoscano

Reviewed By: yurchor, ltoscano

Subscribers: i.Dark_Templar, kde-doc-english

Tags: #documentation

Differential Revision: https://phabricator.kde.org/D16057